refactor: consolidate duplicate type definitions
- Extract Prefecture interface to shared types/prefecture.ts
- Remove duplicate WeatherConditionSchema from ScoreRules.ts
- Update imports to use centralized Prefecture type
- Improve type consistency across domain and usecase layers
