+++ src/types/prefecture.ts
@@ -1,0 +1,7 @@
+export interface Prefecture {
+  id: number;
+  name_ja: string;
+  name_en: string;
+  latitude: number;
+  longitude: number;
+}
