+++ src/usecase/BatchCalculateTouringIndex.ts
@@ -1,15 +1,8 @@
 import type { Weather } from "../domain/Weather";
+import type { Prefecture } from "../types/prefecture";
 import { validateBatchStartDate } from "../utils/dateUtils";
 import { logger } from "../utils/logger";
 import { calculateTouringIndex } from "./CalculateTouringIndex";
-
-export interface Prefecture {
-  id: number;
-  name_ja: string;
-  name_en: string;
-  latitude: number;
-  longitude: number;
-}
 
 export interface TouringIndexBatchItem {
   prefecture_id: number;
